Rename async_api to controller
